+++ core/src/Cuda/Kokkos_Cuda.hpp
@@ -31,7 +31,7 @@ static_assert(false,
 #include <vector>
 
 #include <impl/Kokkos_AnalyzePolicy.hpp>
-#include <Kokkos_CudaSpace.hpp>
+#include <Cuda/Kokkos_CudaSpace.hpp>
 #include <Cuda/Kokkos_Cuda_Error.hpp>  // CUDA_SAFE_CALL
 
 #include <Kokkos_Parallel.hpp>
