+++ core/src/Cuda/Kokkos_Cuda_GraphNode_Impl.hpp
@@ -25,7 +25,7 @@
 
 #include <impl/Kokkos_GraphImpl.hpp>  // GraphAccess needs to be complete
 
-#include <Kokkos_Cuda.hpp>
+#include <Cuda/Kokkos_Cuda.hpp>
 #include <cuda_runtime_api.h>
 
 namespace Kokkos {
